requirements.txt
torch
opencv-python-headless==4.6.0.66
torchvision
Pillow
picozero
pandas
numpy
requests